+++ CHANGES.md
@@ -18,6 +18,9 @@
     uploaded to every superevent and PE is performed for superevents which have
     FAR < 1/(2 weeks).
 
+-   Add a Flask-based web interface for manually triggering certain tasks such
+    as sending updated GCN notices.
+
 ## 0.2.5 (2018-02-01)
 
 -   Pass along the GWCelery version number to Sentry.